[wpiutil] Split function_ref to a separate header
