+++ glass/src/lib/native/include/glass/hardware/AnalogOutput.h
@@ -4,7 +4,7 @@
 
 #pragma once
 
-#include <wpi/STLExtras.h>
+#include <wpi/function_ref.h>
 
 #include "glass/Model.h"
