+++ glass/src/lib/native/include/glass/other/Mechanism2D.h
@@ -7,7 +7,7 @@
 #include <frc/geometry/Rotation2d.h>
 #include <frc/geometry/Translation2d.h>
 #include <imgui.h>
-#include <wpi/STLExtras.h>
+#include <wpi/function_ref.h>
 
 #include "glass/Model.h"
 #include "glass/View.h"
